requirements.txt
altair==5.1.1
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-MySQLdb==1.0.1
Flask-pymysql==0.2.3
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.37
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lit==17.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==8.1.0
mysqlclient @ file:///tmp/build/80754af9/mysqlclient_1609786120714/work
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.1
Pillow==9.5.0
protobuf==4.21.12
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-math==0.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.21
streamlit==1.27.0
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ultralytics==8.0.188
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.17.0
